requirements.txt
typer==0.4.0
digi-xbee==1.4.1
pandas==1.2.1
matplotlib==3.5.1
click==7.1.2

